+++ src/components/todo-tag/TodoTagListView.tsx
@@ -4,7 +4,7 @@ import { useEffect } from "react";
 import { useTodoTagsAllQuery } from "../../api/api-client/Query";
 import { useTags, useTagsDispatch } from "../../contexts/TagsContext";
 import DataTableHeader from "../common/data-tables/DataTableHeader";
-import TodoTagTableRow from "../common/data-tables/TodoTagTable/TodoTagTableRow";
+import TodoTagTableRow from "../common/data-tables/TableRows/TodoTagTableRow";
 
 export function TodoTagListView() {
   const { tags, fetchRequired } = useTags();
